progress on implementing LoadElement
